Move Admin::SwitchUsersController to AdminModeController#edit
