+++ app/views/admin_mode/edit.html.erb
@@ -1,8 +1,11 @@
 <%
-  @title = :app_switch_users.l
+# This is the form for admins to switch users, while in admin mode.
+# It's not in the admin controller because it has different auth
+
+@title = :app_switch_users.l
 %>
 
-<%= form_with(url: admin_switch_users_path, class: "mt-3 pb-2",
+<%= form_with(url: admin_mode_path, class: "mt-3 pb-2", method: :put,
               id: "admin_switch_users_form") do |f| %>
 
   <span class="text-larger"><%= :LOGIN_NAME.t %>:</span><!-- .text-larger --><br/>
