+++ app/views/layouts/sidebar/_admin.html.erb
@@ -3,7 +3,7 @@
 <%= link_to(:app_blocked_ips.t, edit_admin_blocked_ips_path,
             { class: classes[:admin],
               id: "nav_admin_blocked_ips_link" }) %>
-<%= link_to(:app_switch_users.t, new_admin_switch_users_path,
+<%= link_to(:app_switch_users.t, edit_admin_mode_path,
             { class: classes[:admin],
               id: "nav_admin_switch_users_link" }) %>
 <%= link_to(:app_users.t, users_path(by: "name"),
