+++ config/initializers/inflections.rb
@@ -30,5 +30,4 @@ ActiveSupport::Inflector.inflections(:en) do |inflect|
   inflect.acronym("TSV")
   inflect.irregular("bonus", "bonuses")
   inflect.irregular("info", "info")
-  inflect.irregular("account", "account")
 end
