+++ config/routes.rb
@@ -506,12 +506,12 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
   # Route /123 to /observations/123.
   get ":id" => "observations#show", id: /\d+/, as: "permanent_observation"
 
-  # NOTE: Nesting below is necessary to get nice path helpers
-  resource :account, only: [:new, :create]
+  # NOTE: The nesting below is necessary to get nice path helpers
+  resource :account, only: [:new, :create], controller: "account"
 
   namespace :account do
     get("welcome")
-    get("signup", to: "/account#new")
+    get("signup", to: "/account#new") # alternate path
 
     resource :login, only: [:new, :create], controller: "login"
     get("email_new_password", controller: "login")
@@ -523,7 +523,7 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
     get("no_email/:id", to: "preferences#no_email", as: "no_email")
 
     resource :profile, only: [:edit, :update], controller: "profile"
-    patch("profile/remove_image", controller: "profile")
+    patch("profile/remove_image", controller: "profile") # alternate path
 
     resource :verify, only: [:new, :create], controller: "verifications"
     # Alternate path name for email verification
@@ -543,23 +543,18 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
   namespace :admin do
     resource :users, only: [:edit, :update, :destroy]
     resource :donations, only: [:new, :create, :edit, :update, :destroy]
-    get("review_donations", to: "donations#edit")
-    # resource :turn_admin_on, only: [:show], controller: "turn_on"
-    # resource :turn_admin_off, only: [:show], controller: "turn_off"
+    get("review_donations", to: "donations#edit") # alternate path
     resource :banner, only: [:edit, :update], controller: "banner"
-    resource :switch_users, only: [:new, :create]
     resource :blocked_ips, only: [:edit, :update]
     resource :add_user_to_group, only: [:new, :create],
                                  controller: "add_user_to_group"
     namespace :emails do
       resource :feature, only: [:new, :create], controller: "feature"
     end
-    # get("show")
-    get("test_flash_redirection")
   end
 
-  # ----- Admin Mode: just GET with params ---------------------------------
-  get "/admin_mode", to: "admin_mode#show"
+  # ----- Admin Mode: turn on/off, and switch users -----------------------
+  resource :admin_mode, only: [:show, :edit, :update], controller: "admin_mode"
 
   # ----- Articles: standard actions --------------------------------------
   resources :articles, id: /\d+/
@@ -576,7 +571,7 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
   # ----- Contributors: standard actions --------------------------------------
   resources :contributors, only: [:index]
 
-  # ----- Emails: no resources, just forms -------------------------------------
+  # ----- Emails: no resources, just forms ------------------------------------
   match("/emails/ask_observation_question(/:id)",
         to: "emails#ask_observation_question", via: [:get, :post], id: /\d+/,
         as: "emails_ask_observation_question")
@@ -621,11 +616,6 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
   end
   resources :herbaria, id: /\d+/
 
-  # ----- Image:
-  # ----- temporary show route for path_builder with id ---------------
-  get("/image/show_image/:id", to: "image#show_image",
-                               as: "show_image")
-
   # ----- Info: no resources, just forms and pages ----------------------------
   get("/info/how_to_help", to: "info#how_to_help")
   get("/info/how_to_use", to: "info#how_to_use")
@@ -641,15 +631,6 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
   get("/javascript/turn_javascript_off", to: "javascript#turn_javascript_off")
   get("/javascript/turn_javascript_nil", to: "javascript#turn_javascript_nil")
   get("/javascript/hide_thumbnail_map", to: "javascript#hide_thumbnail_map")
-
-  # ----- Location:
-  # ----- temporary show route for path_builder with id ---------------
-  get("/location/show_location/:id", to: "location#show_location",
-                                     as: "show_location")
-
-  # ----- Name:
-  # ----- temporary show route for path_builder with id ---------------
-  get("/name/show_name/:id", to: "name#show_name", as: "show_name")
 
   # ----- Observations: standard actions  ----------------------------
   resources :observations do
@@ -669,21 +650,16 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
   # ----- Policy: one route  --------------------------------------------------
   get("/policy/privacy")
 
-  # ----- Project:
-  # ----- temporary show route for path_builder with id ---------------
-  get("/project/show_project/:id", to: "project#show_project",
-                                   as: "show_project")
-
   # ----- Publications: standard actions  -------------------------------------
   resources :publications
 
-  # ----- RssLogs: nonstandard actions ----------------------------------------
-  # These routes must go before resources, or it will try to match
-  # "rss" to an rss_log
-  # resources :rss_logs, only: [:show, :index]
-  get("/activity_logs/rss", to: "rss_logs#rss", as: "activity_logs_rss")
-  get("/activity_logs", to: "rss_logs#index", as: "activity_logs")
-  get("/activity_logs/:id", to: "rss_logs#show", as: "activity_log")
+  # ----- RssLogs: standard actions ----------------------------------------
+  # This route must go first, or it will try to match "rss" to an rss_log
+  namespace :activity_logs, controller: "rss_logs" do
+    get :rss, to: "/rss_logs#rss"
+  end
+
+  resources :activity_logs, only: [:show, :index], controller: "rss_logs"
 
   # ----- Searches: nonstandard actions --------------------------------------
   match("/search/pattern(/:id)",
@@ -696,11 +672,10 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
   # ----- Sequences: standard actions ---------------------------------------
   resources :sequences, id: /\d+/
 
-  # ----- Species List:
-  # ----- temporary show route for path_builder with id ---------------
-  get("/species_list/show_species_list/:id",
-      to: "species_list#show_species_list",
-      as: "show_species_list")
+  # ----- Test pages  -------------------------------------------
+  namespace :test_pages do
+    resource :flash_redirection, only: [:show], controller: "flash_redirection"
+  end
 
   # ----- Users: standard actions -------------------------------------------
   resources :users, id: /\d+/, only: [:index, :show, :edit, :update]
@@ -709,6 +684,27 @@ MushroomObserver::Application.routes.draw do # rubocop:todo Metrics/BlockLength
   resources :visual_models, id: /\d+/ do
     resources :visual_groups, id: /\d+/, shallow: true
   end
+
+  # Temporary shorter path builders for non-CRUDified controllers SHOW
+
+  # ----- Image:
+  get("/image/show_image/:id", to: "image#show_image",
+                               as: "show_image")
+  # ----- Location:
+  get("/location/show_location/:id", to: "location#show_location",
+                                     as: "show_location")
+  # ----- Name:
+  get("/name/show_name/:id", to: "name#show_name",
+                             as: "show_name")
+  # ----- Project:
+  get("/project/show_project/:id", to: "project#show_project",
+                                   as: "show_project")
+  # ----- Species List:
+  get("/species_list/show_species_list/:id",
+      to: "species_list#show_species_list",
+      as: "show_species_list")
+
+  # ----- end temporary show routes for path_builder with id ---------------
 
   # Short-hand notation for AJAX methods.
   # get "ajax/:action/:type/:id" => "ajax", constraints: { id: /\S.*/ }
