+++ test/controllers/admin_controller_test.rb
@@ -2,22 +2,6 @@
 
 require("test_helper")
 
-# Controller tests for info pages
+# Nothing yet
 class AdminControllerTest < FunctionalTestCase
-  # def test_basic_access
-  #   get(:show)
-  #   assert_redirected_to(herbaria_path)
-
-  #   login
-  #   assert_false(session[:admin])
-  #   rolf.admin = true
-  #   rolf.save!
-
-  #   get(:show)
-  #   assert_redirected_to("/")
-
-  #   session[:admin] = true
-  #   get(:show)
-  #   assert_template("admin/show")
-  # end
 end
