+++ test/controllers/admin_mode_controller_test.rb
@@ -17,4 +17,31 @@ class AdminModeControllerTest < FunctionalTestCase
     get(:show, params: { turn_off: true })
     assert_false(session[:admin])
   end
+
+  def test_switch_users
+    get(:edit)
+    assert_response(:redirect)
+
+    login(:rolf)
+    get(:edit)
+    assert_response(:redirect)
+
+    rolf.admin = true
+    rolf.save!
+    get(:edit)
+    assert_response(:success)
+
+    assert_users_equal(rolf, User.current)
+    put(:update, params: { id: "unverified" })
+    assert_users_equal(rolf, User.current)
+    assert_flash(/not verified yet/)
+    put(:update, params: { id: "Frosted Flake" })
+    assert_users_equal(rolf, User.current)
+    put(:update, params: { id: mary.id })
+    assert_users_equal(mary, User.current)
+    put(:update, params: { id: dick.login })
+    assert_users_equal(dick, User.current)
+    put(:update, params: { id: mary.email })
+    assert_users_equal(mary, User.current)
+  end
 end
